refactor/back-#165: dto 정리
