+++ backend/history-server/src/main/java/com/smiletogether/historyserver/service/dto/request/ChannelMessageDeleteRequest.java
@@ -1,4 +1,4 @@
-package com.smiletogether.historyserver.service.dto;
+package com.smiletogether.historyserver.service.dto.request;
 
 public record ChannelMessageDeleteRequest(
         String type,
