+++ backend/history-server/src/main/java/com/smiletogether/historyserver/service/dto/request/ChannelMessageSaveRequest.java
@@ -1,5 +1,6 @@
-package com.smiletogether.historyserver.service.dto;
+package com.smiletogether.historyserver.service.dto.request;
 
+import com.smiletogether.historyserver.service.dto.WorkspaceProfileDto;
 import java.time.LocalDateTime;
 
 public record ChannelMessageSaveRequest(
