+++ backend/history-server/src/main/java/com/smiletogether/historyserver/service/dto/response/ChannelMessageDeleteResponse.java
@@ -1,4 +1,4 @@
-package com.smiletogether.historyserver.service.dto;
+package com.smiletogether.historyserver.service.dto.response;
 
 public record ChannelMessageDeleteResponse(
         String code,
